fix: allow increasing amount and keeping token expiry the same
This allows users to boost an existing token, while maintaining the same
expiry.

Also changed logic for adjusting cancel time to be more readable. Cancel
time can only be decreased if the token has already expired. Which is
useful for when extending an existing delegation, we don't want users to
be able to decrease the stated cancel time.